articles lib refactor, added new libraries for page and tag functions
