+++ lib/articles.js
@@ -1,25 +1,5 @@
 import { fetchGraphQL } from './graphql';
 
-export const HASURA_LIST_ORG_LOCALES = `query FrontendListOrgLocales {
-  organization_locales {
-    locale {
-      code
-    }
-  }
-}`;
-export function hasuraListLocales(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    orgSlug: params['orgSlug'],
-    query: HASURA_LIST_ORG_LOCALES,
-    name: 'FrontendListOrgLocales',
-    variables: {
-      locale_code: params['localeCode'],
-      slug: params['slug'],
-    },
-  });
-}
-
 const HASURA_SEARCH_ARTICLES = `query FrontendSearchArticles($locale_code: String!, $term: String!) {
   articles(where: {article_translations: {headline: {_ilike: $term}, locale_code: {_eq: $locale_code}, published: {_eq: true}}}) {
     id
@@ -63,213 +43,6 @@ const HASURA_SEARCH_ARTICLES = `query FrontendSearchArticles($locale_code: Strin
         }
         slug
       }
-    }
-  }
-}`;
-
-const HASURA_HOMEPAGE_EDITOR = `query FrontendHomepageEditor($locale_code: String) {
-  homepage_layout_schemas {
-    id
-    name
-    data
-  }
-  homepage_layout_datas {
-    first_article {
-      id
-      article_translations(where: {published: {_eq: true}, locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-        custom_byline
-        facebook_description
-        facebook_title
-        first_published_at
-        headline
-        last_published_at
-        main_image
-        search_description
-        search_title
-        twitter_description
-        twitter_title
-        updated_at
-        locale_code
-      }
-      category {
-        slug
-        id
-        category_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-          locale_code
-          title
-        }
-      }
-      slug
-      author_articles {
-        author {
-          first_names
-          last_name
-          photoUrl
-          slug
-          author_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-            locale_code
-            title
-          }
-        }
-      }
-      tag_articles(where: {tag: {published: {_eq: true}, tag_translations: {locale_code: {_eq: $locale_code}}}}) {
-        tag {
-          tag_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-            locale_code
-            title
-          }
-          slug
-        }
-      }
-    }
-    homepage_layout_schema {
-      id
-      name
-      data
-    }
-    second_article {
-      id
-      article_translations(where: {published: {_eq: true}, locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-        main_image
-        custom_byline
-        facebook_description
-        facebook_title
-        first_published_at
-        headline
-        last_published_at
-        search_description
-        search_title
-        twitter_description
-        twitter_title
-        updated_at
-        locale_code
-      }
-      category {
-        slug
-        id
-        category_translations(order_by: {id: desc}, limit: 1) {
-          locale_code
-          title
-        }
-      }
-      slug
-      author_articles {
-        author {
-          first_names
-          last_name
-          photoUrl
-          slug
-          author_translations(order_by: {id: desc}, limit: 1) {
-            locale_code
-            title
-          }
-        }
-      }
-      tag_articles(where: {tag: {published: {_eq: true}}}) {
-        tag {
-          tag_translations(order_by: {id: desc}, limit: 1) {
-            locale_code
-            title
-          }
-          slug
-        }
-      }
-    }
-    third_article {
-      id
-      article_translations(where: {published: {_eq: true}, locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-        main_image
-        custom_byline
-        facebook_description
-        facebook_title
-        first_published_at
-        headline
-        last_published_at
-        search_description
-        search_title
-        twitter_description
-        twitter_title
-        updated_at
-        locale_code
-      }
-      category {
-        slug
-        id
-        category_translations(order_by: {id: desc}, limit: 1) {
-          locale_code
-          title
-        }
-      }
-      slug
-      author_articles {
-        author {
-          first_names
-          last_name
-          photoUrl
-          slug
-          author_translations(order_by: {id: desc}, limit: 1) {
-            locale_code
-            title
-          }
-        }
-      }
-      tag_articles(where: {tag: {published: {_eq: true}}}) {
-        tag {
-          tag_translations(order_by: {id: desc}, limit: 1) {
-            title
-            locale_code
-          }
-          slug
-        }
-      }
-    }
-  }
-  categories {
-    category_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      title
-      locale_code
-    }
-    published
-    slug
-  }
-  organization_locales {
-    locale {
-      code
-      name
-    }
-  }
-  pages(where: {page_translations: {published: {_eq: true}, locale_code: {_eq: $locale_code}}}) {
-    slug
-    page_translations(where: {published: {_eq: true}, locale_code: {_eq: $locale_code}}, distinct_on: locale_code) {
-      locale_code
-      published
-      headline
-    }
-  }
-  site_metadatas(where: {published: {_eq: true}, site_metadata_translations: {locale_code: {_eq: $locale_code}}}) {
-    site_metadata_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      data
-      locale_code
-    }
-  }
-  tags(where: {published: {_eq: true}, tag_translations: {locale_code: {_eq: $locale_code}}}) {
-    id
-    slug
-    tag_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      title
-      locale_code
-    }
-  }
-}`;
-
-const HASURA_UPSERT_HOMEPAGE = `mutation upsert_homepage_layout_data($homepage_layout_schema_id: Int!, $article_priority_1: Int, $article_priority_2: Int, $article_priority_3: Int) {
-  insert_homepage_layout_datas(objects: [{article_priority_1: $article_priority_1, article_priority_2: $article_priority_2, article_priority_3: $article_priority_3, homepage_layout_schema_id: $homepage_layout_schema_id}], on_conflict: {constraint: homepage_layout_datas_organization_id_key, update_columns: [article_priority_1, article_priority_2, article_priority_3, homepage_layout_schema_id]}) {
-    returning {
-      id
-      homepage_layout_schema_id
-      article_priority_1
-      article_priority_2
-      article_priority_3
     }
   }
 }`;
@@ -448,407 +221,6 @@ const HASURA_ARTICLE_PAGE_SLUG_VERSION = `query FrontendArticlePageSlugVersion($
   }
 }`;
 
-const HASURA_GET_METADATA_BY_LOCALE = `query FrontendGetMetadataByLocale($locale_code: String!) {
-  organization_locales {
-    locale {
-      code
-    }
-  }
-  site_metadatas(where: {site_metadata_translations: {locale_code: {_eq: $locale_code}}}) {
-    id
-    published
-    site_metadata_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      data
-      locale_code
-    }
-  }
-  authors {
-    first_names
-    last_name
-    slug
-  }
-  categories {
-    slug
-    category_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      title
-      locale_code
-    }
-  }
-  pages {
-    slug
-    page_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      headline
-      locale_code
-    }
-  }
-  tags {
-    slug
-    tag_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      title
-      locale_code
-    }
-  }
-}`;
-
-const HASURA_LIST_TAGS_BY_LOCALE = `query FrontendListTagsByLocale($locale_code: String!) {
-  organization_locales {
-    locale {
-      code
-    }
-  }
-  tags(where: {published: {_eq: true}}, order_by: {slug: asc}) {
-    id
-    slug
-    tag_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      locale_code
-      title
-    }
-    tag_articles_aggregate {
-      aggregate {
-        count
-      }
-    }
-  }
-}`;
-
-const HASURA_LIST_TAGS = `query FrontendListTags {
-  organization_locales {
-    locale {
-      code
-    }
-  }
-  tags(where: {published: {_eq: true}}) {
-    id
-    slug
-    tag_translations {
-      locale_code
-      title
-    }
-  }
-}`;
-
-const HASURA_TAG_PAGE = `query FrontendTagPage($tag_slug: String!, $locale_code: String!) {
-  categories {
-    category_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      title
-      locale_code
-    }
-    published
-    slug
-  }
-  organization_locales {
-    locale {
-      code
-      name
-    }
-  }
-  site_metadatas(where: {published: {_eq: true}, site_metadata_translations: {locale_code: {_eq: $locale_code}}}) {
-    site_metadata_translations(order_by: {id: desc}, limit: 1, where: {locale_code: {_eq: $locale_code}}) {
-      data
-      locale_code
-    }
-  }
-  tags(where: {slug: {_eq: $tag_slug}, published: {_eq: true}, tag_translations: {locale_code: {_eq: $locale_code}}}) {
-    id
-    slug
-    tag_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      title
-      locale_code
-    }
-    tag_articles(where: {article: {article_translations: {published: {_eq: true}, locale_code: {_eq: $locale_code}}}}, order_by: {article: {article_translations_aggregate: {min: {first_published_at: desc}}}}) {
-      article {
-        article_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-          custom_byline
-          facebook_description
-          facebook_title
-          first_published_at
-          headline
-          last_published_at
-          main_image
-          search_description
-          search_title
-          twitter_description
-          twitter_title
-          locale_code
-        }
-        category {
-          slug
-          id
-          category_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-            title
-            locale_code
-          }
-        }
-        slug
-        author_articles {
-          author {
-            first_names
-            last_name
-            photoUrl
-            slug
-            author_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-              title
-              locale_code
-            }
-          }
-        }
-      }
-    }
-  }
-}
-`;
-
-const HASURA_CATEGORY_PAGE = `query FrontendCategoryPage($category_slug: String!, $locale_code: String!) {
-  articles(order_by: {article_translations_aggregate: {min: {first_published_at: desc}}}, where: {article_translations: {published: {_eq: true}, locale_code: {_eq: $locale_code}}, category: {slug: {_eq: $category_slug}}}) {
-    article_translations(where: {locale_code: {_eq: $locale_code}, published: {_eq: true}},  order_by: {id: desc}, limit: 1) {
-      custom_byline
-      facebook_description
-      facebook_title
-      first_published_at
-      headline
-      last_published_at
-      main_image
-      search_description
-      search_title
-      twitter_description
-      twitter_title
-      locale_code
-    }
-    category {
-      slug
-      published
-      id
-      category_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-        locale_code
-        title
-      }
-    }
-    slug
-    author_articles {
-      author {
-        first_names
-        last_name
-        photoUrl
-        slug
-        author_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-          locale_code
-          title
-        }
-      }
-    }
-  }
-  categories {
-    category_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      title
-      locale_code
-    }
-    published
-    slug
-  }
-  organization_locales {
-    locale {
-      code
-      name
-    }
-  }
-  site_metadatas(where: {published: {_eq: true}, site_metadata_translations: {locale_code: {_eq: $locale_code}}}) {
-    site_metadata_translations(order_by: {id: desc}, limit: 1, where: {locale_code: {_eq: $locale_code}}) {
-      data
-      locale_code
-    }
-  }
-  tags(where: {published: {_eq: true}}) {
-    slug
-    tag_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      locale_code
-      title
-    }
-  }
-}`;
-
-const HASURA_GET_LAYOUT = `query FrontendGetLayout($locale_code: String!) {
-  categories {
-    slug
-    published
-    category_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      title
-    }
-  }
-  site_metadatas(where: {published: {_eq: true}, site_metadata_translations: {locale_code: {_eq: $locale_code}}}) {
-    site_metadata_translations(order_by: {id: desc}, limit: 1, where: {locale_code: {_eq: $locale_code}}) {
-      data
-    }
-  }
-}`;
-
-const HASURA_GET_PAGE_SLUG_VERSION = `query FrontendGetPageSlugVersion($slug: String!, $locale_code: String!) {
-  page_slug_versions(where: {slug: {_eq: $slug}, page: {page_translations: {published: {_eq: true}, locale_code: {_eq: $locale_code}}}}) {
-    slug
-    page {
-      id
-      author_pages {
-        author {
-          id
-          first_names
-          last_name
-          slug
-          photoUrl
-          author_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-            title
-            locale_code
-          }
-        }
-      }
-      page_translations(where: {published: {_eq: true}, locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-        content
-        facebook_description
-        facebook_title
-        first_published_at
-        headline
-        last_published_at
-        locale_code
-        locale {
-          code
-          name
-        }
-        published
-        search_description
-        search_title
-        twitter_description
-        twitter_title
-      }
-      slug
-    }
-  }
-  authors(order_by: {last_name: asc}) {
-    id
-    first_names
-    last_name
-    created_at
-    slug
-    staff
-    twitter
-    photoUrl
-    published
-    author_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      title
-      bio
-      locale_code
-    }
-  }
-  categories(where: {category_translations: {locale_code: {_eq: $locale_code}}}) {
-    slug
-    published
-    category_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      title
-      locale_code
-    }
-  }
-  organization_locales {
-    locale {
-      code
-      name
-    }
-  }
-  pages(where: {slug: {_eq: $slug}, page_translations: {published: {_eq: true}, locale_code: {_eq: $locale_code}}}) {
-    page_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      locale {
-        code
-        name
-      }
-      id
-    }
-  }
-  site_metadatas(where: {site_metadata_translations: {locale_code: {_eq: $locale_code}}, published: {_eq: true}}) {
-    site_metadata_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      data
-      locale_code
-    }
-  }
-}`;
-
-const HASURA_GET_PAGE_PREVIEW = `query FrontendGetPagePreview($slug: String!, $locale_code: String!) {
-  page_slug_versions(where: {slug: {_eq: $slug}, page: {page_translations: {locale_code: {_eq: $locale_code}}}}) {
-    slug
-    page {
-      id
-      author_pages {
-        author {
-          id
-          first_names
-          last_name
-          slug
-          photoUrl
-          author_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-            title
-          }
-        }
-      }
-      page_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-        content
-        facebook_description
-        facebook_title
-        first_published_at
-        headline
-        last_published_at
-        locale_code
-        locale {
-          code
-          name
-        }
-        published
-        search_description
-        search_title
-        twitter_description
-        twitter_title
-      }
-      slug
-    }
-  }
-  authors(order_by: {last_name: asc}) {
-    id
-    first_names
-    last_name
-    created_at
-    slug
-    staff
-    twitter
-    photoUrl
-    published
-    author_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      title
-      bio
-    }
-  }
-  categories(where: {category_translations: {locale_code: {_eq: $locale_code}}}) {
-    slug
-    published
-    category_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      title
-    }
-  }
-  organization_locales {
-    locale {
-      code
-      name
-    }
-  }
-  pages(where: {slug: {_eq: $slug}}) {
-    page_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      locale {
-        code
-        name
-      }
-      id
-    }
-  }
-  site_metadatas(where: {site_metadata_translations: {locale_code: {_eq: $locale_code}}, published: {_eq: true}}) {
-    site_metadata_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      data
-    }
-  }
-}`;
-
 const HASURA_PREVIEW_ARTICLE_BY_SLUG = `query FrontendPreviewArticleBySlug($slug: String!, $locale_code: String!) {
   articles(where: {slug: {_eq: $slug}, article_translations: {locale_code: {_eq: $locale_code}}}) {
     slug
@@ -927,132 +299,6 @@ const HASURA_GET_ARTICLE_BY_SLUG = `query FrontendGetArticleBySlug($slug: String
   }
 }`;
 
-const HASURA_AUTHOR_PAGE = `query FrontendAuthorPage($author_slug: String!, $locale_code: String!) {
-  articles(order_by: {article_translations_aggregate: {min: {first_published_at: desc}}}, where: {article_translations: {published: {_eq: true}, locale_code: {_eq: $locale_code}}, author_articles: {author: {slug: {_eq: $author_slug}}}}) {
-    article_translations(where: {published: {_eq: true}, locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      custom_byline
-      facebook_description
-      facebook_title
-      first_published_at
-      headline
-      last_published_at
-      main_image
-      search_description
-      search_title
-      twitter_description
-      twitter_title
-      locale_code
-    }
-    category {
-      slug
-      id
-      published
-      category_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-        locale_code
-        title
-      }
-    }
-    slug
-    author_articles(where: {author: {slug: {_eq: $author_slug}}}) {
-      author {
-        first_names
-        last_name
-        photoUrl
-        slug
-        author_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-          locale_code
-          title
-        }
-      }
-    }
-    tag_articles(where: {tag: {published: {_eq: true}}}) {
-      tag {
-        tag_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-          locale_code
-          title
-        }
-        slug
-      }
-    }
-  }
-  categories(where: {published: {_eq: true}, category_translations: {locale_code: {_eq: $locale_code}}}) {
-    category_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      title
-      locale_code
-    }
-    published
-    slug
-  }
-  organization_locales {
-    locale {
-      code
-      name
-    }
-  }
-  site_metadatas(where: {published: {_eq: true}, site_metadata_translations: {locale_code: {_eq: $locale_code}}}) {
-    site_metadata_translations(order_by: {id: desc}, limit: 1, where: {locale_code: {_eq: $locale_code}}) {
-      data
-      locale_code
-    }
-  }
-  authors(where: {slug: {_eq: $author_slug}}, limit: 1) {
-    id
-    slug
-    author_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      bio
-      locale_code
-      title
-    }
-    photoUrl
-    first_names
-    last_name
-    twitter
-    staff
-  }
-}`;
-
-const HASURA_LIST_ALL_SECTIONS = `query FrontendListAllSections {
-  categories {
-    category_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      title
-      locale_code
-    }
-    slug
-  }
-}`;
-
-const HASURA_GET_HOMEPAGE_LAYOUTS = `query FrontendGetHomepageLayouts {
-  homepage_layout_schemas {
-    id
-    name
-    data
-  }
-}`;
-
-const HASURA_GET_SITE_METADATA = `query FrontendGetSiteMetadata($locale_code: String!) {
-  site_metadatas(where: {site_metadata_translations: {locale_code: {_eq: $locale_code}}, published: {_eq: true}}) {
-    site_metadata_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-      data
-      locale_code
-    }
-  }
-}`;
-
-export function hasuraSaveHomepageLayout(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    site: params['site'],
-    query: HASURA_UPSERT_HOMEPAGE,
-    name: 'upsert_homepage_layout_data',
-    variables: {
-      homepage_layout_schema_id: params['schemaId'],
-      article_priority_1: params['article1'],
-      article_priority_2: params['article2'],
-      article_priority_3: params['article3'],
-    },
-  });
-}
-
 export function hasuraSearchArticles(params) {
   let term = '%' + params['term'] + '%';
   let searchVariables = {
@@ -1066,108 +312,6 @@ export function hasuraSearchArticles(params) {
     query: HASURA_SEARCH_ARTICLES,
     name: 'FrontendSearchArticles',
     variables: searchVariables,
-  });
-}
-
-export function hasuraGetHomepageEditor(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    site: params['site'],
-    query: HASURA_HOMEPAGE_EDITOR,
-    name: 'FrontendHomepageEditor',
-    variables: {
-      locale_code: params['localeCode'],
-    },
-  });
-}
-
-export function hasuraGetHomepageLayouts(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    orgSlug: params['orgSlug'],
-    query: HASURA_GET_HOMEPAGE_LAYOUTS,
-    name: 'FrontendGetHomepageLayouts',
-  });
-}
-
-export function hasuraListAllTags(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    orgSlug: params['orgSlug'],
-    query: HASURA_LIST_TAGS,
-    name: 'FrontendListTags',
-  });
-}
-
-export function hasuraListAllTagsByLocale(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    site: params['site'],
-    query: HASURA_LIST_TAGS_BY_LOCALE,
-    name: 'FrontendListTagsByLocale',
-    variables: {
-      locale_code: 'en-US',
-    },
-  });
-}
-
-export function hasuraListAllSections(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    orgSlug: params['orgSlug'],
-    query: HASURA_LIST_ALL_SECTIONS,
-    name: 'FrontendListAllSections',
-  });
-}
-
-export function hasuraGetPagePreview(params) {
-  // console.log('url:', params['url'], 'orgSlug:', params['orgSlug']);
-  return fetchGraphQL({
-    url: params['url'],
-    orgSlug: params['orgSlug'],
-    query: HASURA_GET_PAGE_PREVIEW,
-    name: 'FrontendGetPagePreview',
-    variables: {
-      slug: params['slug'],
-      locale_code: 'en-US',
-    },
-  });
-}
-
-export function hasuraGetPage(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    site: params['site'],
-    query: HASURA_GET_PAGE_SLUG_VERSION,
-    name: 'FrontendGetPageSlugVersion',
-    variables: {
-      locale_code: params['localeCode'],
-      slug: params['slug'],
-    },
-  });
-}
-
-export function hasuraGetLayout(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    site: params['site'],
-    query: HASURA_GET_LAYOUT,
-    name: 'FrontendGetLayout',
-    variables: {
-      locale_code: params['localeCode'],
-    },
-  });
-}
-
-export function hasuraGetMetadataByLocale(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    site: params['site'],
-    query: HASURA_GET_METADATA_BY_LOCALE,
-    name: 'FrontendGetMetadataByLocale',
-    variables: {
-      locale_code: params['localeCode'],
-    },
   });
 }
 
@@ -1195,49 +339,6 @@ export function hasuraGetArticleBySlug(params) {
       slug: params['slug'],
       locale_code: params['localeCode'],
     },
-  });
-}
-
-const HASURA_LIST_PAGE_SLUGS_PREVIEW = `query FrontendListPageSlugsPreview {
-  pages {
-    slug
-    page_translations(distinct_on: locale_code) {
-      locale_code
-      published
-      headline
-    }
-  }
-}`;
-
-const HASURA_LIST_PAGE_SLUGS = `query FrontendListPageSlugs {
-  pages(where: {page_translations: {published: {_eq: true}}}) {
-    slug
-    page_translations(where: {published: {_eq: true}}, distinct_on: locale_code) {
-      locale_code
-      published
-      headline
-    }
-  }
-}`;
-
-export async function hasuraListAllPageSlugsPreview(params = {}) {
-  let url = params['url'] || process.env.HASURA_API_URL;
-  let orgSlug = params['orgSlug'] || process.env.ORG_SLUG;
-  return fetchGraphQL({
-    url: url,
-    orgSlug: orgSlug,
-    query: HASURA_LIST_PAGE_SLUGS_PREVIEW,
-    name: 'FrontendListPageSlugsPreview',
-  });
-}
-export async function hasuraListAllPageSlugs(params = {}) {
-  let url = params['url'] || process.env.HASURA_API_URL;
-  let orgSlug = params['orgSlug'] || process.env.ORG_SLUG;
-  return fetchGraphQL({
-    url: url,
-    orgSlug: orgSlug,
-    query: HASURA_LIST_PAGE_SLUGS,
-    name: 'FrontendListPageSlugs',
   });
 }
 
@@ -1307,55 +408,6 @@ export async function hasuraListAllArticlePreviewSlugs(params = {}) {
   });
 }
 
-const HASURA_GET_AUTHOR_SLUGS = `query FrontendGetAuthorSlugs {
-  authors(where: {published: {_eq: true}}) {
-    slug
-    author_translations {
-      locale_code
-    }
-  }
-}`;
-
-export async function hasuraListAllAuthorPaths(params = {}) {
-  let url = params['url'] || process.env.HASURA_API_URL;
-  let orgSlug = params['orgSlug'] || process.env.ORG_SLUG;
-  return fetchGraphQL({
-    url: url,
-    orgSlug: orgSlug,
-    query: HASURA_GET_AUTHOR_SLUGS,
-    name: 'FrontendGetAuthorSlugs',
-  });
-}
-
-export async function hasuraTagPage(params) {
-  let url = params['url'] || process.env.HASURA_API_URL;
-  return fetchGraphQL({
-    url: url,
-    site: params['site'],
-    query: HASURA_TAG_PAGE,
-    name: 'FrontendTagPage',
-    variables: {
-      locale_code: params['localeCode'],
-      tag_slug: params['tagSlug'],
-    },
-  });
-}
-
-export async function hasuraCategoryPage(params) {
-  let url = params['url'] || process.env.HASURA_API_URL;
-
-  return fetchGraphQL({
-    url: url,
-    site: params['site'],
-    query: HASURA_CATEGORY_PAGE,
-    name: 'FrontendCategoryPage',
-    variables: {
-      locale_code: params['localeCode'],
-      category_slug: params['categorySlug'],
-    },
-  });
-}
-
 export async function hasuraPreviewArticlePage(params) {
   let url = params['url'] || process.env.HASURA_API_URL;
 
@@ -1382,33 +434,6 @@ export async function hasuraArticlePage(params) {
     variables: {
       category_slug: params['categorySlug'],
       slug: params['slug'],
-      locale_code: params['localeCode'],
-    },
-  });
-}
-
-export async function hasuraAuthorPage(params) {
-  let url = params['url'] || process.env.HASURA_API_URL;
-
-  return fetchGraphQL({
-    url: url,
-    site: params['site'],
-    query: HASURA_AUTHOR_PAGE,
-    name: 'FrontendAuthorPage',
-    variables: {
-      author_slug: params['authorSlug'],
-      locale_code: params['localeCode'],
-    },
-  });
-}
-
-export async function hasuraGetSiteMetadata(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    site: params['site'],
-    query: HASURA_GET_SITE_METADATA,
-    name: 'FrontendGetSiteMetadata',
-    variables: {
       locale_code: params['localeCode'],
     },
   });
@@ -1661,77 +686,6 @@ export async function hasuraUpdateArticle(params) {
       url: params['url'],
       created_by_email: params['created_by_email'],
       article_sources: params['article_sources'],
-    },
-  });
-}
-
-const HASURA_CREATE_PAGE = `mutation FrontendTestingCreatePage($slug: String!, $locale_code: String!, $created_by_email: String, $document_id: String, $url: String, $facebook_title: String, $facebook_description: String, $search_title: String, $search_description: String, $headline: String, $twitter_title: String, $twitter_description: String, $content: jsonb, $published: Boolean) {
-  insert_pages(
-    objects: {
-      page_google_documents: {
-        data: {
-          google_document: {
-            data: {
-              document_id: $document_id, locale_code: $locale_code, url: $url
-            },
-            on_conflict: {
-              constraint: google_documents_organization_id_document_id_key,
-              update_columns: [document_id]
-            }
-          }
-        },
-        on_conflict: {
-          constraint: page_google_documents_page_id_google_document_id_key,
-          update_columns: [google_document_id]
-        }
-      },
-      slug: $slug,
-      page_translations: {
-        data: {
-          created_by_email: $created_by_email, published: $published, search_description: $search_description, search_title: $search_title, twitter_description: $twitter_description, twitter_title: $twitter_title, locale_code: $locale_code, headline: $headline, facebook_title: $facebook_title, facebook_description: $facebook_description, content: $content
-        }
-      }
-    },
-    on_conflict: {
-      constraint: pages_slug_organization_id_key,
-      update_columns: [slug, updated_at]
-    }) {
-    returning {
-      id
-      slug
-      page_google_documents {
-        id
-        google_document {
-          document_id
-          locale_code
-          url
-        }
-      }
-      page_translations(where: { locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
-        published
-      }
-    }
-  }
-}`;
-
-export async function hasuraCreatePage(params) {
-  let url = params['url'] || process.env.HASURA_API_URL;
-  let orgSlug = params['orgSlug'] || process.env.ORG_SLUG;
-  return fetchGraphQL({
-    url: url,
-    orgSlug: orgSlug,
-    query: HASURA_CREATE_PAGE,
-    name: 'FrontendTestingCreatePage',
-    variables: {
-      published: params['published'],
-      slug: params['slug'],
-      locale_code: params['locale_code'],
-      headline: params['headline'],
-      content: params['content'],
-      search_title: params['search_title'],
-      search_description: params['search_description'],
-      document_id: params['document_id'],
-      created_by_email: params['created_by_email'],
     },
   });
 }
@@ -2114,111 +1068,6 @@ export async function hasuraGetGoogleDocsForPage(params) {
   });
 }
 
-const HASURA_LIST_DOMAINS = `query FrontendListCustomDomainsAndSubdomains {
-  organizations(where: {_or: [{subdomain: {_is_null: false}}, {customDomain: {_is_null: false}}]}) {
-    customDomain
-    subdomain
-    organization_locales {
-      locale {
-        code
-      }
-    }
-  }
-}`;
-
-export async function listDomains(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    adminSecret: params['adminSecret'],
-    query: HASURA_LIST_DOMAINS,
-    name: 'FrontendListCustomDomainsAndSubdomains',
-  });
-}
-
-export async function generateAllDomainPaths(params) {
-  const { errors, data } = await listDomains(params);
-  if (errors) {
-    throw errors;
-  }
-
-  let allPaths = [];
-  for (let i = 0; i < data.organizations.length; i++) {
-    let org = data.organizations[i];
-    let siteParams = Object.assign({}, params['urlParams']);
-    if (org.subdomain) {
-      siteParams['site'] = org.subdomain;
-    } else if (org.customDomain) {
-      siteParams['site'] = org.customDomain;
-    }
-
-    // trying without locale - debugging 404 en-US Vercel Platforms rollout
-    allPaths.push({
-      params: siteParams,
-    });
-  }
-
-  return allPaths;
-}
-
-const HASURA_LIST_CATEGORY_PAGE_PATHS = `query FrontendListCategoryPagePaths {
-  organizations(where: {_or: [{subdomain: {_is_null: false}}, {customDomain: {_is_null: false}}]}) {
-    customDomain
-    subdomain
-    organization_locales {
-      locale {
-        code
-      }
-    }
-    categories(where: {published: {_eq: true}}) {
-      id
-      slug
-      category_translations {
-        locale_code
-        title
-      }
-      
-    }
-  }
-}`;
-
-export async function listCategoryPagePaths(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    adminSecret: params['adminSecret'],
-    query: HASURA_LIST_CATEGORY_PAGE_PATHS,
-    name: 'FrontendListCategoryPagePaths',
-  });
-}
-
-const HASURA_LIST_TAG_PAGE_PATHS = `query FrontendListTagPagePaths {
-  organizations(where: {_or: [{subdomain: {_is_null: false}}, {customDomain: {_is_null: false}}]}) {
-    customDomain
-    subdomain
-    organization_locales {
-      locale {
-        code
-      }
-    }
-    tags(where: {published: {_eq: true}}) {
-      id
-      slug
-      tag_translations {
-        locale_code
-        title
-      }
-    }
-  }
-}`;
-
-export async function listTagPagePaths(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    adminSecret: params['adminSecret'],
-    query: HASURA_LIST_TAG_PAGE_PATHS,
-    name: 'FrontendListTagPagePaths',
-  });
-}
-
 const HASURA_LIST_STATIC_PREVIEW_PAGE_PATHS = `query FrontendListStaticPreviewPagePaths {
   organizations(where: {_or: [{subdomain: {_is_null: false}}, {customDomain: {_is_null: false}}]}) {
     customDomain
@@ -2317,178 +1166,6 @@ export async function generateAllStaticPagePaths(params = {}) {
   return allPaths;
 }
 
-export async function generateAllTagPagePaths(params = {}) {
-  const { errors, data } = await listTagPagePaths(params);
-  if (errors) {
-    throw errors;
-  }
-
-  let allPaths = [];
-  for (let i = 0; i < data.organizations.length; i++) {
-    let org = data.organizations[i];
-
-    for (let x = 0; x < org.tags.length; x++) {
-      let tag = org.tags[x];
-      let siteParams = Object.assign({}, params['urlParams']);
-      if (org.subdomain) {
-        siteParams['site'] = org.subdomain;
-      } else if (org.customDomain) {
-        siteParams['site'] = org.customDomain;
-      }
-      siteParams['slug'] = tag.slug;
-      allPaths.push({
-        params: siteParams,
-      });
-    }
-  }
-
-  return allPaths;
-}
-
-export async function generateAllCategoryPagePaths(params = {}) {
-  const { errors, data } = await listCategoryPagePaths(params);
-  if (errors) {
-    throw errors;
-  }
-
-  let allPaths = [];
-  for (let i = 0; i < data.organizations.length; i++) {
-    let org = data.organizations[i];
-
-    for (let x = 0; x < org.categories.length; x++) {
-      let category = org.categories[x];
-      let siteParams = Object.assign({}, params['urlParams']);
-      if (org.subdomain) {
-        siteParams['site'] = org.subdomain;
-      } else if (org.customDomain) {
-        siteParams['site'] = org.customDomain;
-      }
-      siteParams['category'] = category.slug;
-      allPaths.push({
-        params: siteParams,
-      });
-    }
-  }
-
-  return allPaths;
-}
-
-const HASURA_LIST_AUTHOR_PAGE_PATHS = `query FrontendListAuthorPagePaths {
-  organizations(where: {_or: [{subdomain: {_is_null: false}}, {customDomain: {_is_null: false}}]}) {
-    customDomain
-    subdomain
-    organization_locales {
-      locale {
-        code
-      }
-    }
-    authors(where: {published: {_eq: true}}) {
-      id
-      slug
-      author_translations {
-        locale_code
-        title
-      }
-    }
-  }
-}`;
-
-export async function listAuthorPagePaths(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    adminSecret: params['adminSecret'],
-    query: HASURA_LIST_AUTHOR_PAGE_PATHS,
-    name: 'FrontendListAuthorPagePaths',
-  });
-}
-
-export async function generateAllAuthorPagePaths(params = {}) {
-  const { errors, data } = await listAuthorPagePaths(params);
-  if (errors) {
-    throw errors;
-  }
-
-  let allPaths = [];
-  for (let i = 0; i < data.organizations.length; i++) {
-    let org = data.organizations[i];
-
-    for (let x = 0; x < org.authors.length; x++) {
-      let author = org.authors[x];
-      let siteParams = Object.assign({}, params['urlParams']);
-      if (org.subdomain) {
-        siteParams['site'] = org.subdomain;
-      } else if (org.customDomain) {
-        siteParams['site'] = org.customDomain;
-      }
-      siteParams['slug'] = author.slug;
-
-      allPaths.push({
-        params: siteParams,
-      });
-    }
-  }
-
-  return allPaths;
-}
-
-const HASURA_LIST_NEWSLETTER_PAGE_PATHS = `query FrontendListNewsletterPagePaths {
-  organizations(where: {_or: [{subdomain: {_is_null: false}}, {customDomain: {_is_null: false}}]}) {
-    customDomain
-    subdomain
-    organization_locales {
-      locale {
-        code
-      }
-    }
-    newsletter_editions(order_by: {newsletter_published_at: desc}) {
-      byline
-      content
-      headline
-      newsletter_published_at
-      slug
-      subheadline
-    }
-  }
-}`;
-
-export async function listNewsletterPagePaths(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    adminSecret: params['adminSecret'],
-    query: HASURA_LIST_NEWSLETTER_PAGE_PATHS,
-    name: 'FrontendListNewsletterPagePaths',
-  });
-}
-
-export async function generateAllNewsletterPagePaths(params = {}) {
-  const { errors, data } = await listNewsletterPagePaths(params);
-  if (errors) {
-    console.error('generateAllNewsletterPagePaths errors:', errors);
-    throw errors;
-  }
-
-  let allPaths = [];
-  for (let i = 0; i < data.organizations.length; i++) {
-    let org = data.organizations[i];
-
-    for (let x = 0; x < org.newsletter_editions.length; x++) {
-      let newsletter = org.newsletter_editions[x];
-      let siteParams = Object.assign({}, params['urlParams']);
-      if (org.subdomain) {
-        siteParams['site'] = org.subdomain;
-      } else if (org.customDomain) {
-        siteParams['site'] = org.customDomain;
-      }
-      siteParams['slug'] = newsletter.slug;
-      allPaths.push({
-        params: siteParams,
-      });
-    }
-  }
-
-  return allPaths;
-}
-
 const HASURA_LIST_ALL_ARTICLE_ARCHIVE_PAGES = `query FrontendPaginatedListArticlesArchivePages {
   organizations(where: {_or: [{subdomain: {_is_null: false}}, {customDomain: {_is_null: false}}]}) {
     customDomain
@@ -2575,7 +1252,7 @@ export async function generateAllArticlePreviewPagePaths(params = {}) {
 export async function generateAllArticleArchivePages(params = {}) {
   const { errors, data } = await listArticleArchivePages(params);
   if (errors) {
-    console.error('generateAllNewsletterPagePaths errors:', errors);
+    console.error('errors:', errors);
     throw errors;
   }
