+++ lib/authors.js
@@ -114,6 +114,129 @@ const HASURA_UPDATE_AUTHOR = `mutation updateAuthorWithTranslations($author_id: 
   }
 }`;
 
+const HASURA_AUTHOR_PAGE = `query FrontendAuthorPage($author_slug: String!, $locale_code: String!) {
+  articles(order_by: {article_translations_aggregate: {min: {first_published_at: desc}}}, where: {article_translations: {published: {_eq: true}, locale_code: {_eq: $locale_code}}, author_articles: {author: {slug: {_eq: $author_slug}}}}) {
+    article_translations(where: {published: {_eq: true}, locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
+      custom_byline
+      facebook_description
+      facebook_title
+      first_published_at
+      headline
+      last_published_at
+      main_image
+      search_description
+      search_title
+      twitter_description
+      twitter_title
+      locale_code
+    }
+    category {
+      slug
+      id
+      published
+      category_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
+        locale_code
+        title
+      }
+    }
+    slug
+    author_articles(where: {author: {slug: {_eq: $author_slug}}}) {
+      author {
+        first_names
+        last_name
+        photoUrl
+        slug
+        author_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
+          locale_code
+          title
+        }
+      }
+    }
+    tag_articles(where: {tag: {published: {_eq: true}}}) {
+      tag {
+        tag_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
+          locale_code
+          title
+        }
+        slug
+      }
+    }
+  }
+  categories(where: {published: {_eq: true}, category_translations: {locale_code: {_eq: $locale_code}}}) {
+    category_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
+      title
+      locale_code
+    }
+    published
+    slug
+  }
+  organization_locales {
+    locale {
+      code
+      name
+    }
+  }
+  site_metadatas(where: {published: {_eq: true}, site_metadata_translations: {locale_code: {_eq: $locale_code}}}) {
+    site_metadata_translations(order_by: {id: desc}, limit: 1, where: {locale_code: {_eq: $locale_code}}) {
+      data
+      locale_code
+    }
+  }
+  authors(where: {slug: {_eq: $author_slug}}, limit: 1) {
+    id
+    slug
+    author_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
+      bio
+      locale_code
+      title
+    }
+    photoUrl
+    first_names
+    last_name
+    twitter
+    staff
+  }
+}`;
+const HASURA_LIST_AUTHOR_PAGE_PATHS = `query FrontendListAuthorPagePaths {
+  organizations(where: {_or: [{subdomain: {_is_null: false}}, {customDomain: {_is_null: false}}]}) {
+    customDomain
+    subdomain
+    organization_locales {
+      locale {
+        code
+      }
+    }
+    authors(where: {published: {_eq: true}}) {
+      id
+      slug
+      author_translations {
+        locale_code
+        title
+      }
+    }
+  }
+}`;
+
+const HASURA_GET_AUTHOR_SLUGS = `query FrontendGetAuthorSlugs {
+  authors(where: {published: {_eq: true}}) {
+    slug
+    author_translations {
+      locale_code
+    }
+  }
+}`;
+
+export async function hasuraListAllAuthorPaths(params = {}) {
+  let url = params['url'] || process.env.HASURA_API_URL;
+  let orgSlug = params['orgSlug'] || process.env.ORG_SLUG;
+  return fetchGraphQL({
+    url: url,
+    orgSlug: orgSlug,
+    query: HASURA_GET_AUTHOR_SLUGS,
+    name: 'FrontendGetAuthorSlugs',
+  });
+}
+
 export function hasuraListAllAuthors(params = {}) {
   let url = params['url'] || process.env.HASURA_API_URL;
   // console.log('listAllAuthors locale:', localeCode);
@@ -218,4 +341,57 @@ export function deleteSingleAuthor(params) {
       id: params['id'],
     },
   });
+}
+
+export async function hasuraAuthorPage(params) {
+  let url = params['url'] || process.env.HASURA_API_URL;
+
+  return fetchGraphQL({
+    url: url,
+    site: params['site'],
+    query: HASURA_AUTHOR_PAGE,
+    name: 'FrontendAuthorPage',
+    variables: {
+      author_slug: params['authorSlug'],
+      locale_code: params['localeCode'],
+    },
+  });
+}
+
+export async function listAuthorPagePaths(params) {
+  return fetchGraphQL({
+    url: params['url'],
+    adminSecret: params['adminSecret'],
+    query: HASURA_LIST_AUTHOR_PAGE_PATHS,
+    name: 'FrontendListAuthorPagePaths',
+  });
+}
+
+export async function generateAllAuthorPagePaths(params = {}) {
+  const { errors, data } = await listAuthorPagePaths(params);
+  if (errors) {
+    throw errors;
+  }
+
+  let allPaths = [];
+  for (let i = 0; i < data.organizations.length; i++) {
+    let org = data.organizations[i];
+
+    for (let x = 0; x < org.authors.length; x++) {
+      let author = org.authors[x];
+      let siteParams = Object.assign({}, params['urlParams']);
+      if (org.subdomain) {
+        siteParams['site'] = org.subdomain;
+      } else if (org.customDomain) {
+        siteParams['site'] = org.customDomain;
+      }
+      siteParams['slug'] = author.slug;
+
+      allPaths.push({
+        params: siteParams,
+      });
+    }
+  }
+
+  return allPaths;
 }
