+++ lib/newsletters.js
@@ -88,3 +88,61 @@ export function hasuraGetNewsletter(params) {
     },
   });
 }
+
+const HASURA_LIST_NEWSLETTER_PAGE_PATHS = `query FrontendListNewsletterPagePaths {
+  organizations(where: {_or: [{subdomain: {_is_null: false}}, {customDomain: {_is_null: false}}]}) {
+    customDomain
+    subdomain
+    organization_locales {
+      locale {
+        code
+      }
+    }
+    newsletter_editions(order_by: {newsletter_published_at: desc}) {
+      byline
+      content
+      headline
+      newsletter_published_at
+      slug
+      subheadline
+    }
+  }
+}`;
+
+export async function listNewsletterPagePaths(params) {
+  return fetchGraphQL({
+    url: params['url'],
+    adminSecret: params['adminSecret'],
+    query: HASURA_LIST_NEWSLETTER_PAGE_PATHS,
+    name: 'FrontendListNewsletterPagePaths',
+  });
+}
+
+export async function generateAllNewsletterPagePaths(params = {}) {
+  const { errors, data } = await listNewsletterPagePaths(params);
+  if (errors) {
+    console.error('generateAllNewsletterPagePaths errors:', errors);
+    throw errors;
+  }
+
+  let allPaths = [];
+  for (let i = 0; i < data.organizations.length; i++) {
+    let org = data.organizations[i];
+
+    for (let x = 0; x < org.newsletter_editions.length; x++) {
+      let newsletter = org.newsletter_editions[x];
+      let siteParams = Object.assign({}, params['urlParams']);
+      if (org.subdomain) {
+        siteParams['site'] = org.subdomain;
+      } else if (org.customDomain) {
+        siteParams['site'] = org.customDomain;
+      }
+      siteParams['slug'] = newsletter.slug;
+      allPaths.push({
+        params: siteParams,
+      });
+    }
+  }
+
+  return allPaths;
+}
