+++ lib/section.js
@@ -1,19 +1,71 @@
 import { fetchGraphQL } from './graphql';
 import _ from 'lodash';
 
-const HASURA_GET_TAG_BY_ID = `query FrontendGetTagByID($id: Int!) {
+const HASURA_CATEGORY_PAGE = `query FrontendCategoryPage($category_slug: String!, $locale_code: String!) {
+  articles(order_by: {article_translations_aggregate: {min: {first_published_at: desc}}}, where: {article_translations: {published: {_eq: true}, locale_code: {_eq: $locale_code}}, category: {slug: {_eq: $category_slug}}}) {
+    article_translations(where: {locale_code: {_eq: $locale_code}, published: {_eq: true}},  order_by: {id: desc}, limit: 1) {
+      custom_byline
+      facebook_description
+      facebook_title
+      first_published_at
+      headline
+      last_published_at
+      main_image
+      search_description
+      search_title
+      twitter_description
+      twitter_title
+      locale_code
+    }
+    category {
+      slug
+      published
+      id
+      category_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
+        locale_code
+        title
+      }
+    }
+    slug
+    author_articles {
+      author {
+        first_names
+        last_name
+        photoUrl
+        slug
+        author_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
+          locale_code
+          title
+        }
+      }
+    }
+  }
+  categories {
+    category_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
+      title
+      locale_code
+    }
+    published
+    slug
+  }
   organization_locales {
     locale {
       code
+      name
     }
   }
-  tags_by_pk(id: $id) {
-    id
-    published
-    tag_translations {
+  site_metadatas(where: {published: {_eq: true}, site_metadata_translations: {locale_code: {_eq: $locale_code}}}) {
+    site_metadata_translations(order_by: {id: desc}, limit: 1, where: {locale_code: {_eq: $locale_code}}) {
+      data
+      locale_code
+    }
+  }
+  tags(where: {published: {_eq: true}}) {
+    slug
+    tag_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
+      locale_code
       title
     }
-    slug
   }
 }`;
 
@@ -38,6 +90,15 @@ const HASURA_LIST_ALL_SECTIONS_BY_LOCALE = `query FrontendListAllSectionsByLocal
     }
   }
 }`;
+const HASURA_LIST_ALL_SECTIONS = `query FrontendListAllSections {
+  categories {
+    category_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
+      title
+      locale_code
+    }
+    slug
+  }
+}`;
 
 const HASURA_UPSERT_SECTION = `mutation FrontendUpsertSection($locale_code: String!, $title: String!, $slug: String!, $published: Boolean) {
   insert_categories(objects: {slug: $slug, category_translations: {data: {locale_code: $locale_code, title: $title}, on_conflict: {constraint: category_translations_locale_code_category_id_key, update_columns: [title]}}, published: $published}, on_conflict: {constraint: categories_organization_id_slug_key, update_columns: [slug, published]}) {
@@ -45,17 +106,6 @@ const HASURA_UPSERT_SECTION = `mutation FrontendUpsertSection($locale_code: Stri
       id
       slug
       published
-    }
-  }
-}`;
-
-const HASURA_INSERT_TAG = `mutation insert_single_tag($locale_code: String, $title: String, $published: Boolean, $slug: String) {
-  insert_tags_one(object: {slug: $slug, published: $published, tag_translations: {data: {locale_code: $locale_code, title: $title}}}) {
-    id
-    slug
-    tag_translations {
-      locale_code
-      title
     }
   }
 }`;
@@ -101,19 +151,6 @@ const HASURA_GET_SECTION_BY_ID = `query FrontendGetSectionByID($id: Int!, $local
   }
 }`;
 
-const HASURA_UPDATE_TAG = `mutation updateTagWithTranslations($tag_id: Int!, $locale_code: String!, $title: String, $published: Boolean, $slug: String) {
-  delete_tag_translations(where: {tag_id: {_eq: $tag_id}, locale_code: {_eq: $locale_code}}) {
-    affected_rows
-  }
-  insert_tag_translations(objects: [{tag_id: $tag_id, locale_code: $locale_code, title: $title}]) {
-    affected_rows
-  }
-  update_tags_by_pk(pk_columns: {id: $tag_id}, _set: {published: $published, slug: $slug}) {
-    published
-    slug
-  }
-}`;
-
 const HASURA_UPDATE_SECTION = `mutation updateCategoryWithTranslations($category_id: Int!, $locale_code: String!, $title: String, $published: Boolean, $slug: String) {
   delete_category_translations(where: {category_id: {_eq: $category_id}, locale_code: {_eq: $locale_code}}) {
     affected_rows
@@ -126,7 +163,86 @@ const HASURA_UPDATE_SECTION = `mutation updateCategoryWithTranslations($category
     slug
   }
 }`;
+const HASURA_LIST_CATEGORY_PAGE_PATHS = `query FrontendListCategoryPagePaths {
+  organizations(where: {_or: [{subdomain: {_is_null: false}}, {customDomain: {_is_null: false}}]}) {
+    customDomain
+    subdomain
+    organization_locales {
+      locale {
+        code
+      }
+    }
+    categories(where: {published: {_eq: true}}) {
+      id
+      slug
+      category_translations {
+        locale_code
+        title
+      }
+      
+    }
+  }
+}`;
+export async function hasuraCategoryPage(params) {
+  let url = params['url'] || process.env.HASURA_API_URL;
 
+  return fetchGraphQL({
+    url: url,
+    site: params['site'],
+    query: HASURA_CATEGORY_PAGE,
+    name: 'FrontendCategoryPage',
+    variables: {
+      locale_code: params['localeCode'],
+      category_slug: params['categorySlug'],
+    },
+  });
+}
+
+export async function listCategoryPagePaths(params) {
+  return fetchGraphQL({
+    url: params['url'],
+    adminSecret: params['adminSecret'],
+    query: HASURA_LIST_CATEGORY_PAGE_PATHS,
+    name: 'FrontendListCategoryPagePaths',
+  });
+}
+
+export function hasuraListAllSections(params) {
+  return fetchGraphQL({
+    url: params['url'],
+    orgSlug: params['orgSlug'],
+    query: HASURA_LIST_ALL_SECTIONS,
+    name: 'FrontendListAllSections',
+  });
+}
+
+export async function generateAllCategoryPagePaths(params = {}) {
+  const { errors, data } = await listCategoryPagePaths(params);
+  if (errors) {
+    throw errors;
+  }
+
+  let allPaths = [];
+  for (let i = 0; i < data.organizations.length; i++) {
+    let org = data.organizations[i];
+
+    for (let x = 0; x < org.categories.length; x++) {
+      let category = org.categories[x];
+      let siteParams = Object.assign({}, params['urlParams']);
+      if (org.subdomain) {
+        siteParams['site'] = org.subdomain;
+      } else if (org.customDomain) {
+        siteParams['site'] = org.customDomain;
+      }
+      siteParams['category'] = category.slug;
+      allPaths.push({
+        params: siteParams,
+      });
+    }
+  }
+
+  return allPaths;
+}
 export function hasuraListAllSectionsByLocale(params) {
   return fetchGraphQL({
     url: params['url'],
@@ -154,45 +270,6 @@ export function hasuraUpsertSection(params) {
   });
 }
 
-const DELETE_TAG_MUTATION = `mutation DeleteSingleTag($id: Int!) {
-  delete_tag_translations(where: {tag_id: {_eq: $id}}) {
-    affected_rows
-  }
-  delete_tag_articles(where: {tag_id: {_eq: $id}}) {
-    affected_rows
-  }
-  delete_tags(where: {id: {_eq: $id}}) {
-    affected_rows
-  }
-}`;
-
-export function deleteSingleTag(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    site: params['site'],
-    query: DELETE_TAG_MUTATION,
-    name: 'DeleteSingleTag',
-    variables: {
-      id: params['id'],
-    },
-  });
-}
-
-export function hasuraCreateTag(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    site: params['site'],
-    query: HASURA_INSERT_TAG,
-    name: 'insert_single_tag',
-    variables: {
-      locale_code: params['localeCode'],
-      title: params['title'],
-      published: params['published'],
-      slug: params['slug'],
-    },
-  });
-}
-
 export function hasuraCreateSection(params) {
   return fetchGraphQL({
     url: params['url'],
@@ -208,18 +285,6 @@ export function hasuraCreateSection(params) {
   });
 }
 
-export function hasuraGetTagById(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    site: params['site'],
-    query: HASURA_GET_TAG_BY_ID,
-    name: 'FrontendGetTagByID',
-    variables: {
-      id: params['id'],
-    },
-  });
-}
-
 export function hasuraGetSectionById(params) {
   return fetchGraphQL({
     url: params['url'],
@@ -229,22 +294,6 @@ export function hasuraGetSectionById(params) {
     variables: {
       id: params['id'],
       localeCode: params['localeCode'],
-    },
-  });
-}
-
-export function hasuraUpdateTag(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    site: params['site'],
-    query: HASURA_UPDATE_TAG,
-    name: 'updateTagWithTranslations',
-    variables: {
-      tag_id: params['id'],
-      locale_code: params['localeCode'],
-      title: params['title'],
-      published: params['published'],
-      slug: params['slug'],
     },
   });
 }
