+++ lib/site_metadata.js
@@ -23,3 +23,78 @@ export function hasuraUpsertMetadata(params) {
     },
   });
 }
+
+const HASURA_GET_METADATA_BY_LOCALE = `query FrontendGetMetadataByLocale($locale_code: String!) {
+  organization_locales {
+    locale {
+      code
+    }
+  }
+  site_metadatas(where: {site_metadata_translations: {locale_code: {_eq: $locale_code}}}) {
+    id
+    published
+    site_metadata_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
+      data
+      locale_code
+    }
+  }
+  authors {
+    first_names
+    last_name
+    slug
+  }
+  categories {
+    slug
+    category_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
+      title
+      locale_code
+    }
+  }
+  pages {
+    slug
+    page_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
+      headline
+      locale_code
+    }
+  }
+  tags {
+    slug
+    tag_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
+      title
+      locale_code
+    }
+  }
+}`;
+
+export function hasuraGetMetadataByLocale(params) {
+  return fetchGraphQL({
+    url: params['url'],
+    site: params['site'],
+    query: HASURA_GET_METADATA_BY_LOCALE,
+    name: 'FrontendGetMetadataByLocale',
+    variables: {
+      locale_code: params['localeCode'],
+    },
+  });
+}
+
+const HASURA_GET_SITE_METADATA = `query FrontendGetSiteMetadata($locale_code: String!) {
+  site_metadatas(where: {site_metadata_translations: {locale_code: {_eq: $locale_code}}, published: {_eq: true}}) {
+    site_metadata_translations(where: {locale_code: {_eq: $locale_code}}, order_by: {id: desc}, limit: 1) {
+      data
+      locale_code
+    }
+  }
+}`;
+
+export async function hasuraGetSiteMetadata(params) {
+  return fetchGraphQL({
+    url: params['url'],
+    site: params['site'],
+    query: HASURA_GET_SITE_METADATA,
+    name: 'FrontendGetSiteMetadata',
+    variables: {
+      locale_code: params['localeCode'],
+    },
+  });
+}
