+++ pages/_sites/[site]/about.js
@@ -1,8 +1,6 @@
 import { useRouter } from 'next/router';
-import {
-  generateAllDomainPaths,
-  hasuraGetPage,
-} from '../../../lib/articles.js';
+import { generateAllDomainPaths } from '../../../lib/settings';
+import { hasuraGetPage } from '../../../lib/pages.js';
 import AboutPage from '../../../components/AboutPage';
 
 export default function About(props) {
