+++ pages/_sites/[site]/preview/thank-you.js
@@ -1,9 +1,8 @@
 import tw from 'twin.macro';
 import { useRouter } from 'next/router';
-import { hasuraGetPage } from '../../../../lib/articles.js';
+import { hasuraGetPage } from '../../../../lib/pages.js';
 import { useAnalytics } from '../../../../lib/hooks/useAnalytics.js';
 import { renderBody } from '../../../../lib/utils';
-import ReadInOtherLanguage from '../../../../components/articles/ReadInOtherLanguage';
 import Layout from '../../../../components/Layout';
 import NewsletterBlock from '../../../../components/plugins/NewsletterBlock';
 import {
