+++ pages/_sites/[site]/tinycms/nav/index.js
@@ -3,7 +3,7 @@ import { useRouter } from 'next/router';
 import tw, { styled } from 'twin.macro';
 import { DragDropContext, Droppable, Draggable } from 'react-beautiful-dnd';
 import { SelectorIcon } from '@heroicons/react/solid';
-import { hasuraGetMetadataByLocale } from '../../../../../lib/articles.js';
+import { hasuraGetMetadataByLocale } from '../../../../../lib/site_metadata';
 import { findSetting, getOrgSettings } from '../../../../../lib/settings.js';
 import AdminLayout from '../../../../../components/AdminLayout.js';
 import AdminNav from '../../../../../components/nav/AdminNav';
