+++ script/populate.js
@@ -1,12 +1,12 @@
 #! /usr/bin/env node
 
-require('dotenv').config({ path: '.env.local' })
+require('dotenv').config({ path: '.env.local' });
 
 const fs = require('fs');
 const path = require('path');
 const fetch = require('node-fetch');
 
-const shared = require("./shared");
+const shared = require('./shared');
 
 const apiUrl = process.env.HASURA_API_URL;
 const apiToken = process.env.ORG_SLUG;
@@ -17,52 +17,6 @@ function writeCache(name, data) {
   fs.writeFileSync(cachedFile, JSON.stringify(data), { flag: 'w' }, (err) => {
     console.log('failed to write file:', err);
   });
-}
-
-async function listLocales() {
-  const localeResult = await shared.hasuraListLocales({
-    url: apiUrl,
-    orgSlug: apiToken,
-  });
-
-  let locales;
-  if (localeResult.errors) {
-    console.error("Error listing locales:", localeResult.errors);
-  } else {
-    locales = localeResult.data.organization_locales;
-  }
-
-  writeCache('locales', locales);
-}
-
-async function listSections() {
-  const result = await shared.hasuraListSections({
-    url: apiUrl,
-    orgSlug: apiToken,
-  });
-
-  let sections;
-  if (result.errors) {
-    console.error("Error listing sections:", result.errors);
-  } else {
-    sections = result.data.categories;
-  }
-  writeCache('sections', sections);
-}
-
-async function listTags() {
-  const result = await shared.hasuraListTags({
-    url: apiUrl,
-    orgSlug: apiToken,
-  });
-
-  let tags;
-  if (result.errors) {
-    console.error("Error listing tags:", result.errors);
-  } else {
-    tags = result.data.tags;
-  }
-  writeCache('tags', tags);
 }
 
 function getAds() {
@@ -89,10 +43,6 @@ function getAds() {
 }
 
 async function main() {
-  console.log("HASURA_API_URL:", process.env.HASURA_API_URL, "ORG_SLUG:", process.env.ORG_SLUG);
-  listLocales();
-  listSections();
-  listTags();
   getAds();
 }
